Fixed a few warnings.
Rev: src/cpp.c:1.74
Rev: src/preprocessor.h:1.32

+++ src/preprocessor.h
@@ -1,5 +1,5 @@
 /*
- * $Id: preprocessor.h,v 1.31 2000/08/10 15:00:20 grubba Exp $
+ * $Id: preprocessor.h,v 1.32 2000/08/10 15:01:01 grubba Exp $
  *
  * Preprocessor template.
  * Based on cpp.c 1.45
@@ -1879,7 +1879,7 @@ static ptrdiff_t lower_cpp(struct cpp *this,
 	/* NOTE: Reuses undefine_ for undef_ */
       if(WGOBBLE2(undefine_) || WGOBBLE2(undef_))
 	{
-	  INT32 tmp;
+	  ptrdiff_t tmp;
 	  struct pike_string *s;
 
 	  SKIPSPACE();
